backend/requirements.txt
requests
Flask
Pyrebase4
flask-cors
tabula-py
pandas
Flask-APScheduler
sendgrid